Add tests and GitHub Actions
